✨ introduce 'significance' concept to the system
Now vector clouds have a 'significance', which is determined by weighing words that are unique to the class more heavily than words that occur frequently across different classes.

This caused several changes, such as the change of IntVec.kt to Vector.kt, which now contains doubles instead of integers for arithmetic convenience, in particular to allow vectors to be normalized to unit vectors (i.e. vectors can now contain not only absolute word count, but also, instead, the relative frequency of a word within the article).