🚀 rework loading logic for release
Later, when released - this application will be packaged inside a JAR file, meaning that the file loading logic had to be reworked in order for everything to work as it was supposed to.

This commit in particular, has to be tested on a different machine as well, since my environment includes things like fonts and resources, which *could* be fallen back to.