✅ provide full data coverage
Added more tests to (almost) fully cover all different cases for data loading.

The only missing cases are OS dependent and thus not 'testable' in a single run (that aside, the macOS-part still crashes due to uncertainty whether the OS query for it is correct (perhaps it's Darwin instead of Mac).